Trekk ut hjelpefunksjonar for sortering til eiga fil slik at det blir mindre info å forhalde seg til i brukartabellen
