+++ src/components/user-table/user-table.tsx
@@ -1,10 +1,14 @@
 import { SortState, Table } from '@navikt/ds-react';
 import { UserTableHeader } from './header/user-table-header';
 import { UserTableBody } from './body/user-table-body';
-import { OrderByDirection, OrderByField } from '../../rest/api';
+import { OrderByField } from '../../rest/api';
 import { useSokStore } from '../../stores/sok-store';
-import { OrNothing } from '../../utils/types/ornothing';
 import './user-table.css';
+import {
+	finnNesteSorteringsretning,
+	finnNesteSorteringsverdi,
+	mapOrderByDirectionToAkselSortDirection
+} from './sortering-utils';
 
 interface ScopedSortState extends SortState {
 	orderBy: OrderByField;
@@ -20,40 +24,13 @@ export const UserTable = () => {
 			}
 		: undefined;
 
-	const nextDir = (
-		nyKey: OrderByField,
-		gamalKey: OrNothing<OrderByField>,
-		gamalDir: OrNothing<OrderByDirection>
-	): OrderByDirection | undefined => {
-		if (gamalKey && gamalDir && nyKey === gamalKey) {
-			if (gamalDir === OrderByDirection.ASC) {
-				return OrderByDirection.DESC;
-			}
-			if (gamalDir === OrderByDirection.DESC) {
-				return undefined;
-			}
-		}
-		return OrderByDirection.ASC;
-	};
-
-	const nextOrderByField = (
-		nyKey: OrderByField,
-		gamalKey: OrNothing<OrderByField>,
-		gamalDir: OrNothing<OrderByDirection>
-	): OrderByField | undefined => {
-		if (nyKey === gamalKey && gamalDir === OrderByDirection.DESC) {
-			return undefined;
-		}
-		return nyKey;
-	};
-
 	// Vekslar mellom "valgt kolonne asc", "valgt kolonne desc" og "ingen valgt kolonne" på kvart tredje klikk på same kolonneoverskrift.
 	const handleSort = (sortKey: ScopedSortState['orderBy']) => {
 		const oldSortKey = orderByField;
 		const oldSortDir = orderByDirection;
 
-		setOrderByField(nextOrderByField(sortKey, oldSortKey, oldSortDir));
-		setOrderByDirection(nextDir(sortKey, oldSortKey, oldSortDir));
+		setOrderByField(finnNesteSorteringsverdi(sortKey, oldSortKey, oldSortDir));
+		setOrderByDirection(finnNesteSorteringsretning(sortKey, oldSortKey, oldSortDir));
 	};
 
 	return (
@@ -70,17 +47,4 @@ export const UserTable = () => {
 			</Table>
 		</div>
 	);
-};
-
-type AkselSortDirection = 'ascending' | 'descending' | 'none';
-
-const mapOrderByDirectionToAkselSortDirection = (dir: OrNothing<OrderByDirection>): AkselSortDirection => {
-	switch (dir) {
-		case OrderByDirection.ASC:
-			return 'ascending';
-		case OrderByDirection.DESC:
-			return 'descending';
-		default:
-			return 'none';
-	}
 };
